Stupid PHP and it's stupid associative vs indexed arrays
