+++ src/Enum/EnumInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bolt\Enum;
+
+interface EnumInterface
+{
+    public static function all(): array;
+
+    public static function isValid(?string $status): bool;
+}
